requirements.txt
absl-py==2.1.0
aiohttp==3.9.4
aioquic==1.2.0
aiosignal==1.3.1
altair==4.2.0
annotated-types==0.6.0
ansi2html==1.9.1
anthropic==0.25.1
anyio==4.3.0
anywidget==0.9.13
appdirs==1.4.4
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
autocommand==2.2.2
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.35.73
botocore==1.35.73
bqplot==0.12.43
branca==0.7.2
Brotli==1.1.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cheroot==10.0.1
CherryPy==18.9.0
CIRpy==1.0.2
click==8.1.7
cloudpickle==3.0.0
colour==0.1.5
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.3
cssselect==1.2.0
curl_cffi==0.6.2
cycler==0.12.1
Cython==3.0.11
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
Django==5.0
dm-tree==0.1.8
duckdb==1.1.3
duckduckgo_search==5.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.1
extra-streamlit-components==0.1.71
faerun==0.4.7
fake-useragent==1.5.1
Faker==30.1.0
fastjsonschema==2.19.1
favicon==0.7.0
filelock==3.13.4
Flask==3.0.3
flatbuffers==24.3.25
folium==0.17.0
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
fuzzywuzzy==0.18.0
gast==0.6.0
gdown==5.2.0
geographiclib==2.0
geojson==3.1.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.67.1
h11==0.14.0
h2==4.1.0
h5py==3.12.1
hpack==4.0.0
htbuilder==0.6.2
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hydralit_components==1.0.10
hyperframe==6.0.1
idna==3.6
importlib_metadata==7.1.0
inflect==7.2.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.29.4
ipyleaflet==0.19.2
ipython==8.23.0
ipytree==0.2.2
ipyvue==1.11.2
ipyvuetify==1.10.0
ipywidgets==8.1.2
itables==2.2.1
itsdangerous==2.2.0
jaraco.collections==5.0.1
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpickle==3.2.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-dash==0.4.2
jupyter-leaflet==0.19.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaitaistruct==0.10
keras==3.6.0
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-text-splitters==0.0.1
langsmith==0.1.47
ldap3==2.9.1
leafmap==0.41.2
libclang==18.1.1
llvmlite==0.42.0
loguru==0.7.2
lxml==5.3.0
lxml_html_clean==0.2.2
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mitmproxy==11.0.0
mitmproxy-macos==0.9.3
mitmproxy_rs==0.9.3
ml-dtypes==0.4.1
molplotly==1.1.5
more-itertools==10.2.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==1.17.1
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
parse==1.20.2
parso==0.8.3
passlib==1.7.4
patsy==1.0.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
playwright==1.47.0
plotly==5.20.0
pmdarima==2.0.4
portend==3.2.0
prettytable==3.12.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
protobuf==5.28.2
psutil==5.9.8
psygnal==0.11.1
ptyprocess==0.7.0
PubChemPy==1.0.4
publicsuffix2==2.20191221
pure-eval==0.2.2
py3Dmol==2.1.0
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pyecharts==2.0.7
pyee==12.0.0
Pygments==2.17.2
PyJWT==2.8.0
pylsqpack==0.3.18
pymdown-extensions==10.11.2
Pympler==1.0.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.9.0
pyppeteer==2.0.0
pyquery==2.0.1
pyshp==2.3.1
PySocks==1.7.1
pystac==1.11.0
pystac-client==0.8.5
python-box==7.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.0
pytz==2024.1
pyvis==0.3.2
PyYAML==6.0.1
pyzmq==26.0.0
rdkit==2023.9.5
referencing==0.34.0
requests==2.32.3
requests-html==0.10.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3fs==0.4.2
s3transfer==0.10.4
schedule==1.2.2
scikit-base==0.11.0
scikit-learn==1.4.1.post1
scipy==1.13.0
scooby==0.10.0
selenium==4.22.0
semver==3.0.2
service-identity==24.1.0
setuptools==70.0.0
shap==0.45.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sktime==0.34.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.29
sqlparse==0.4.4
st-annotated-text==4.0.1
st-material-table==0.0.4
st-mui-table==0.0.4
st-pages==0.4.5
st-theme==1.2.3
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.40.1
streamlit-aggrid==1.0.1
streamlit-authenticator==0.1.5
streamlit-awesome-table==0.1.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.7
streamlit-faker==0.0.3
streamlit-feedback==0.1.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.3.13
streamlit-pdf-viewer==0.0.18
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_folium==0.22.1
stvis==0.0.2
tempora==5.5.1
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-probability==0.25.0
termcolor==2.5.0
threadpoolctl==3.4.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.2
traittypes==0.2.1
trio==0.25.1
trio-websocket==0.11.1
trubrics==1.6.2
typeguard==4.2.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
urwid==2.6.15
validators==0.28.0
w3lib==2.2.1
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wheel==0.45.0
whitebox==2.3.5
whiteboxgui==2.3.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.2
xyzservices==2024.9.0
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.18.1
zstandard==0.23.0
